+++ poi-scratchpad/src/main/java/org/apache/poi/hwmf/record/HwmfEscape.java
@@ -260,8 +260,8 @@ public class HwmfEscape implements HwmfRecord {
         @Override
         public int init(LittleEndianInputStream leis, long recordSize, EscapeFunction escapeFunction) throws IOException {
             this.escapeFunction = escapeFunction;
-            escapeDataBytes = IOUtils.toByteArray(leis,(int)recordSize,MAX_OBJECT_SIZE);
-            return (int)recordSize;
+            escapeDataBytes = IOUtils.toByteArray(leis, recordSize, MAX_OBJECT_SIZE);
+            return Math.toIntExact(recordSize);
         }
 
         @Override
@@ -303,9 +303,9 @@ public class HwmfEscape implements HwmfRecord {
             if (commentIdentifier != EMF_COMMENT_IDENTIFIER) {
                 // there are some WMF implementation using this record as a MFCOMMENT or similar
                 // if the commentIdentifier doesn't match, then return immediately
-                emfData = IOUtils.toByteArray(leis, (int)(recordSize-LittleEndianConsts.INT_SIZE), MAX_OBJECT_SIZE);
+                emfData = IOUtils.toByteArray(leis, recordSize-LittleEndianConsts.INT_SIZE, MAX_OBJECT_SIZE);
                 remainingBytes = emfData.length;
-                return (int)recordSize;
+                return Math.toIntExact(recordSize);
             }
 
             // A 32-bit unsigned integer that identifies the type of comment in this record.
